backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
pandas>=2.2.0
openpyxl>=3.1.0
google-generativeai>=0.8.0
resend>=2.19.0
slowapi>=0.1.9
email-validator>=2.2.0
